OGP取得のエンドポイントを作成
